Remove useless arguments and separate large objects
