+++ Logger.ts
@@ -1,0 +1,18 @@
+import winston, { format } from "winston";
+
+export default (): winston.Logger => {
+  return winston.createLogger({
+    level: "info",
+    format: format.combine(
+      format.colorize(),
+      format.printf(({ level, message }) => { return `${level} ${message}`; })
+    ),
+    transports: [
+      new winston.transports.Console(),
+      new winston.transports.File({
+        level: "debug",
+        filename: "app.log"
+      })
+    ],
+  });
+};
